update travis.yml, add mocha options, restructure test folder
